VisionCore: Split stoneFinder to multiple files.
